Generation in one file, drawing in another
